Refactores nabar item components
They are now less clutter and more isolated. It means they are standalone components, are more modular
