+++ src/components/Navbar.jsx
@@ -1,8 +1,8 @@
 import MyLogo from '../assets/images/VentoTech-icon.png';
-import NavbarFavicon from './navabar-components/navbar-favicon';
 import NavbarItemBar from './navabar-components/navbar-items-bar';
 import BurgerButton from './navabar-components/navbar-burger-button';
 import { BugerButtonIcon } from './Icon-components/Icons-svg';
+import { NavbarFavicon } from './navabar-components/navbar-items-bar-components';
 
 /* 
     This is the final navbar displayed in my portfolio web.
